# way faster constructor