--- a/src/componentarray.jl
+++ b/src/componentarray.jl
@@ -138,12 +138,18 @@ make_carray_args(::NamedTuple{(), Tuple{}}) = (Any[], FlatAxis())
 make_carray_args(::Type{T}, ::NamedTuple{(), Tuple{}}) where {T} = (T[], FlatAxis())
 function make_carray_args(nt)
     data, ax = make_carray_args(Vector, nt)
-    data = length(data)==1 ? [data[1]] : reduce(vcat, data)
+    data = length(data)==1 ? [data[1]] : eltype(data) == Any ? reduce(vcat, data) : data
     return (data, ax)
 end
 make_carray_args(::Type{T}, nt) where {T} = make_carray_args(Vector{T}, nt)
 function make_carray_args(A::Type{<:AbstractArray}, nt)
-    data, idx = make_idx([], nt, 0)
+    init = try
+        T = recursive_type(nt)
+        isa(T, Type) ? T[] : []
+    catch
+        []
+    end
+    data, idx = make_idx(init, nt, 0)
     return (A(data), Axis(idx))
 end