--- a/src/utils.jl
+++ b/src/utils.jl
@@ -48,5 +48,4 @@ recursive_eltype(nt::NamedTuple) = mapreduce(recursive_eltype, promote_type, nt)
 recursive_eltype(x::Vector) = mapreduce(recursive_eltype, promote_type, x)
 recursive_eltype(x::Dict) = mapreduce(recursive_eltype, promote_type, values(x))
 recursive_eltype(::AbstractArray{T,N})  where {T<:Number, N}= T
-recursive_eltype(x::Number) = typeof(x)
-recursive_eltype(x) = Base.eltypeof(x)
+recursive_eltype(x) = typeof(x)