Maybe I misunderstood, `reduce` looks like it's for restoring the type. If the type is determined at the beginning, then there is no need to call `reduce`.

```julia
julia> data_1 = rand(30,20);

julia> data_2 = rand(30,20);

julia> @benchmark ComponentArray(a = $data_1, b = $data_2) # Original
BenchmarkTools.Trial: 8075 samples with 1 evaluation.
 Range (min … max):  372.000 μs …   5.988 ms  ┊ GC (min … max):  0.00% … 90.97%
 Time  (median):     451.300 μs               ┊ GC (median):     0.00%
 Time  (mean ± σ):   615.631 μs ± 819.858 μs  ┊ GC (mean ± σ):  25.07% ± 16.55%

  ██▄▃▁                                                     ▁ ▁ ▂
  █████▇▆▅▃▅▄▄▄▁▃▁▃▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▅████ █
  372 μs        Histogram: log(frequency) by time          5 ms <

 Memory estimate: 2.96 MiB, allocs estimate: 2442.

julia>  @benchmark ComponentArray(a = $data_1, b = $data_2) # This PR
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   9.400 μs …  1.983 ms  ┊ GC (min … max): 0.00% … 98.03%
 Time  (median):     11.100 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   13.263 μs ± 49.256 μs  ┊ GC (mean ± σ):  9.68% ±  2.60%

     ▅▆▇▇█▆▆▅▅▂▂▁▁▂▁▂▃▂▂▂▃▁▂▁▂▁▁                              ▂
  ▂▃███████████████████████████████▇▇▇▅▆▇▇▆▆▆▇▆▆▆▆▃▄▅▅▅▅▄▆▅▄▅ █
  9.4 μs       Histogram: log(frequency) by time      22.7 μs <

 Memory estimate: 32.77 KiB, allocs estimate: 43.
```