@MilkshakeForReal 2022-07-31
The current method of creating a component array from a nested dict is a little strange, I think it's better to convert the dict into a named tuple first.