@MilkshakeForReal 2022-07-31
New benchmark
```julia
julia> @benchmark ComponentArray(a = $data_1, b = $data_2)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   9.600 μs …  4.390 ms  ┊ GC (min … max):  0.00% … 99.16%
 Time  (median):     15.950 μs              ┊ GC (median):     0.00%
 Time  (mean ± σ):   20.259 μs ± 97.878 μs  ┊ GC (mean ± σ):  13.10% ±  2.95%

  ▂▇█▅▅▅▅▇▇▇▆▄▃▁▁           ▁▁▁▁  ▁▁▁▂▁                       ▂
  ███████████████████▇█▇▇▇████████████████▆▅▅▄▄▄▃▅▅▅▇▆▆▆▅▇▆▅▆ █
  9.6 μs       Histogram: log(frequency) by time      59.6 μs <

 Memory estimate: 42.25 KiB, allocs estimate: 43.
```