@avik-pal 2022-08-06
This change seems to have broken https://github.com/avik-pal/Lux.jl/runs/7702517164?check_suite_focus=true. @jonniedie could we add a downstream test dep on Lux similar to https://github.com/avik-pal/Lux.jl/blob/main/.github/workflows/Downstream.yml?