@MilkshakeForReal 2022-08-06
I will do it.

[This line](https://github.com/avik-pal/Lux.jl/blob/main/src/utils.jl#L145) is very domain specific @avik-pal, it could cause strange behaviour. Not so sure what is the motivation of it.

```julia
julia> ComponentArray(a=Int[], b=(;))
ComponentVector{Float32}(a = Float32[], b = Float32[])

julia> ComponentArray(a=Int[1], b=(;))
ComponentVector{Int64}(a = [1], b = Int64[])
```