@avik-pal 2022-08-06
I think at some point it would lead to a `ComponentVector{Any}` which is why I had that hacky thing. We should be able to remove that without any trouble