@MilkshakeForReal 2022-08-06
Not sure if this [PR](https://github.com/jonniedie/ComponentArrays.jl/pull/162) would satisfy your requirement