@avik-pal 2022-08-06
Yeah it should do it https://github.com/avik-pal/Lux.jl/pull/126